+++ docs/source/changelog.rst
@@ -16,6 +16,7 @@ Version (dev)
   allow reporting the source location in exceptions.
 * :py:meth:`Keyvalues.export() <srctools.keyvalues.Keyvalues.export>` is now deprecated.
 * Allow directly passing enums to set VMF keyvalues and fixups, if the ``value`` is itself a valid value.
+* Parse Strata Source's other VMF additions - viewport configuration, brush face vertices and instance visibility.
 * Add :py:attr:`~srctools.tokenizer.Tokenizer.plus_operator`, allowing `+` to be
   parsed as an operator for FGDs but still be valid inside bare strings elsewhere.
   These are common in ``gameinfo.txt``.